collector/requirements.txt
azure-eventhub
azure-storage-blob
pandas
pyarrow
